Modified validator size to be `u64`

+++ chain-config/src/lib.rs
@@ -16,8 +16,8 @@ pub trait ChainConfigContract:
     #[init]
     fn init(
         &self,
-        min_validators: usize,
-        max_validators: usize,
+        min_validators: u64,
+        max_validators: u64,
         min_stake: BigUint,
         admin: ManagedAddress,
         additional_stake_required: MultiValueEncoded<StakeMultiArg<Self::Api>>,
